requirements.txt
python==3.7.13
pytorch==1.11.0
pytorch-pretrained-bert==0.6.2 
torchaudio==0.11.0
torchvision==0.12.0
transformers==4.9.0
tokenizers==0.10.3
pandas==1.3.5
streamlit==1.9.2 

